Update configure.ac to add blas and lapack flags
--with-blas and --with-lapack can now be used to build against a user-
specified lib{blas,lapack}.{a,so,dylib}, and the necessary flags should
be available as $BLAS_LIBS and $LAPACK_LIBS for use in Makefile.am.

These configure.ac features require the use of the "autoconf-archive", a
user-contributed collection of autoconf macros. Added autoconf-archive
to the conda environment*.yml files, as well as blas and liblapack
dependencies, which will be used in functions still to be ported.